moved trefleAPI  and apiMethods to plants folder
